got server up w/o error
